+++ tests/test_i8cfail_0002.py
@@ -1,0 +1,40 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2015 Red Hat, Inc.
+# This file is part of the Infinity Note Compiler.
+#
+# The Infinity Note Compiler is free software: you can redistribute it
+# and/or modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation, either version 3 of
+# the License, or (at your option) any later version.
+#
+# The Infinity Note Compiler is distributed in the hope that it will
+# be useful, but WITHOUT ANY WARRANTY; without even the implied
+# warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+# See the GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with the Infinity Note Compiler.  If not, see
+# <http://www.gnu.org/licenses/>.
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+from tests import TestCase
+from i8c.compiler import StackError
+
+# Stack underflow checks were off by one.
+
+SOURCE = """\
+define test::i8cfail_0002 returns int
+    argument int x
+    load 5
+    blt label
+label:
+"""
+
+class TestI8CFail0002(TestCase):
+    def test_i8cfail_0002(self):
+        """Miscellaneous I8C failure #0002 check"""
+        self.assertRaises(StackError, self.compile, SOURCE)